backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
pillow
numpy
coremltools
